Restore keys.py
