+++ coreblocks/params/__init__.py
@@ -3,3 +3,4 @@ from .optypes import *  # noqa: F401
 from .genparams import *  # noqa: F401
 from .layouts import *  # noqa: F401
 from .fu_params import *  # noqa: F401
+from .keys import *  # noqa: F401
